Adds "Scroll and delete" interaction mode
This commit introduces a new "Scroll and delete" interaction mode as an alternative to the default "Swipe to choose" mode.

In the new mode, gestures are remapped:
*   Swipe left to delete.
*   Swipe right to share.
*   Swipe up/down to navigate to the next/previous photo with a vertical slide animation.

The interaction mode can be changed in the Settings screen. The language selection UI in Settings has also been moved into a dialog to accommodate the new option.